libobs: Fix compiler warnings
CMake build framework 3.0 introduces more strict typing checks, which
require additional code fixes to silence.

* Fixes warnings about implicit casts
* Fixes warnings about non-exhaustive switch statements
* Fixes warnings about variables shadowing pre-existing variables
* Fixes warnings about unused assignments

+++ libobs/obs-output.c
@@ -1230,7 +1230,7 @@ static bool add_caption(struct obs_output *output, struct encoder_packet *out)
 			}
 
 			uint16_t captionData = ((uint8_t *)caption_buf)[1];
-			captionData = captionData << 8;
+			captionData = (uint16_t)(captionData << 8);
 			captionData += ((uint8_t *)caption_buf)[2];
 
 			// padding
